fix empty unsub/punsub TypeError
When unsubscribe or punsubscribe is called and it has
no subscriptions, the reply[1] is a null which causes
`TypeError: Cannot call method 'toString' of null`

Failing tests are provided along with fix.

Fix is to check if reply[1] is null before calling toString otherwise
just pass null.

Stack trace for exception:

```
TypeError: Cannot call method 'toString' of null
    at RedisClient.return_reply (/Users/barczewskij/projects/node_redis/index.js:633:65)
    at ReplyParser.RedisClient.init_parser (/Users/barczewskij/projects/node_redis/index.js:266:14
    at ReplyParser.EventEmitter.emit (events.js:96:17)
    at ReplyParser.send_reply (/Users/barczewskij/projects/node_redis/lib/parser/javascript.js:300
    at ReplyParser.execute (/Users/barczewskij/projects/node_redis/lib/parser/javascript.js:211:22
    at RedisClient.on_data (/Users/barczewskij/projects/node_redis/index.js:483:27)
    at Socket.<anonymous> (/Users/barczewskij/projects/node_redis/index.js:82:14)
    at Socket.EventEmitter.emit (events.js:96:17)
    at TCP.onread (net.js:396:14)
```